\name{delete_rule}
\alias{delete_rule}
\title{Deletes the given rule form the parser.}
\usage{
delete_rule(genE, rule.id)
}
\arguments{
  \item{parser,}{a peg parser produced by new.parser}

  \item{rule.id,}{a character string naming the rule}
}
\description{
Deletes the given rule form the parser.
}

