\name{generate}
\alias{generate}
\title{Generate a pegR parser}
\usage{
generate(grammer = text, testGrammer = FALSE, pegE = NULL)
}
\arguments{
  \item{grammer}{text of the PEG format used to generate
  the resulting parser}

  \item{textGrammer}{}
}
\value{
an enviroment containing the Rule Objects
}
\description{
This function generates a PEG parser
}
\keyword{PEG}
\keyword{grammer}
\keyword{parser}

