\name{get_description}
\alias{get_description}
\title{Gets a description of a given rule}
\usage{
get_description(genE, rule.id)
}
\arguments{
  \item{rule.id,}{a character string naming the rule}

  \item{parser,}{a peg parser produced by new.parser}
}
\value{
description, a character string describing the parser
}
\description{
Gets a description of a given rule
}

