\name{paste1}
\alias{paste1}
\title{A convenient shortcut for pasting lists}
\usage{
paste1(v)
}
\arguments{
  \item{v,}{a list}
}
\description{
This is equivalent to list(paste(v,collapse=''). It is
analogous to pastes0 for vectors.
}
\references{
a list of length one, having its sole element the result of
pasting together the contents of v without any seperator
}

