\name{plot.PEGResult}
\alias{plot.PEGResult}
\title{Plot the parsing tree}
\usage{
\method{plot}{PEGResult}(res, bg = ifelse(match(par("bg"), "transparent", 0),
  "white", par("bg")), border = TRUE, xpad = 1, ypad = 2.5, cex = 1,
  shadow = TRUE, adj = 0.5, ...)
}
\description{
Plot the parsing tree
}

