\name{add_rule}
\alias{add_rule}
\title{add_rule
Adds a rule to the parser}
\usage{
add_rule(parser, rule)
}
\arguments{
  \item{parser,}{a peg parser produced by new.parser}

  \item{rule,}{a quoted string that defines a rule
  according to the PEG Grammer}
}
\value{
Status and the rule processed
}
\description{
add_rule Adds a rule to the parser
}

