\name{set_description}
\alias{set_description}
\title{Attaches an (optional) description to the given rule.}
\usage{
set_description(genE, rule.id, description)
}
\arguments{
  \item{parser,}{a peg parser produced by new.parser}

  \item{rule.id,}{a character string naming the rule}

  \item{description,}{a text string describing the rule}
}
\description{
A description should be used to comment a given rule
}

